backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy==2.0.25
psycopg2-binary
openai==1.76.0
requests==2.31.0
beautifulsoup4==4.12.3
alembic==1.14.1
PyPDF2==3.0.1
python-multipart==0.0.9

google-genai
redis==5.0.4
rq==1.16.2
PyJWT==2.10.1
